// src/types/types.ts
export type HeroType = {
    img: {
        responsiveImage: {
            width: number;
            height: number;
            base64: string;
            src: string;
        };
    };
};

export type TestType = {
    title: string;
    desc: string;
    img: {
        responsiveImage: {
            width: number;
            height: number;
            base64: string;
            src: string;
        };
    };
};